Refactor: 인증 등록 페이지 SSR로 변경 작업 시작
