+++ src/app/(providers)/(root)/challenges/[id]/verification/register/_components/registerVerification/index.ts
@@ -1,0 +1,1 @@
+export { default } from './VerificationRegister';
